move QuadraticBezierTo to separate file
