+++ core/src/commonMain/kotlin/dev/lennartegb/vec2compose/core/commands/Command.kt
@@ -8,17 +8,6 @@ sealed interface Command {
     val isAbsolute: Boolean
 
     /**
-     * Quadratic bezier to command indicated by q/Q
-     */
-    data class QuadraticBezierTo(
-        val x1: Float,
-        val y1: Float,
-        val x2: Float,
-        val y2: Float,
-        override val isAbsolute: Boolean
-    ) : Command
-
-    /**
      * Reflective quadratic bezier to command indicated by t/T
      */
     data class ReflectiveQuadraticBezierTo(
